refactor: 重构interface_info表,替换原本DO
