--- api-core/src/main/java/com/dhx/apicore/mapper/InterfaceExampleEntityMapper.java
@@ -1,17 +1,0 @@
-package com.dhx.apicore.mapper;
-
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-
-/**
-* @author dhx
-* @description 针对表【t_interface_example_entity】的数据库操作Mapper
-* @createDate 2023-04-12 09:38:35
-* @Entity generator.domain.InterfaceExampleEntity
-*/
-public interface InterfaceExampleEntityMapper extends BaseMapper<InterfaceExampleEntity> {
-
-}
-
-
-
-
