+++ api-core/src/main/java/com/dhx/apicore/mapper/InterfaceInfoEntityMapper.java
@@ -1,0 +1,18 @@
+package com.dhx.apicore.mapper;
+
+import com.dhx.apicore.model.DO.InterfaceInfoEntity;
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+
+/**
+* @author dhx
+* @description 针对表【interface_info】的数据库操作Mapper
+* @createDate 2023-12-26 00:34:47
+* @Entity com.dhx.apicore.model.DO.InterfaceInfo
+*/
+public interface InterfaceInfoEntityMapper extends BaseMapper<InterfaceInfoEntity> {
+
+}
+
+
+
+
