--- api-core/src/main/java/com/dhx/apicore/model/DO/InterfaceEntity.java
@@ -1,189 +1,0 @@
-package com.dhx.apicore.model.DO;
-
-import com.baomidou.mybatisplus.annotation.IdType;
-import com.baomidou.mybatisplus.annotation.TableField;
-import com.baomidou.mybatisplus.annotation.TableId;
-import com.baomidou.mybatisplus.annotation.TableName;
-import java.io.Serializable;
-import java.util.Date;
-import lombok.Data;
-
-/**
- *
- * @TableName t_interface_entity
- */
-@TableName(value ="t_interface")
-@Data
-public class InterfaceEntity implements Serializable {
-    /**
-     * 主键
-     */
-    @TableId(type = IdType.AUTO)
-    private Long id;
-
-    /**
-     * 接口名称
-     */
-    private String name;
-
-    /**
-     * 接口描述
-     */
-    private String description;
-
-    /**
-     * 接口调用次数
-     */
-    private Long callTimes;
-
-    /**
-     * 是否免费
-     */
-    private Integer isFree;
-
-    /**
-     * 接口地址
-     */
-    private String url;
-
-    /**
-     * 接口标签
-     */
-    private String tag;
-
-    /**
-     * 接口背景图片
-     */
-    private String background;
-
-    /**
-     * 请求方式
-     */
-    private String method;
-
-    /**
-     * 请求参数
-     */
-    private String requestParam;
-
-    /**
-     * 请求头
-     */
-    private String requestHeader;
-
-    /**
-     * 响应头
-     */
-    private String responseHeader;
-
-    /**
-     * 接口状态(0-关闭 ,1-开启)
-     */
-    private Integer status;
-
-    /**
-     * 创建人
-     */
-    private Integer userId;
-
-    /**
-     * 创建时间
-     */
-    private Date createTime;
-
-    /**
-     * 更新时间
-     */
-    private Date updateTime;
-
-    /**
-     * 逻辑删除(1删除)
-     */
-    private Integer isDelete;
-
-    @TableField(exist = false)
-    private static final long serialVersionUID = 1L;
-
-    @Override
-    public boolean equals(Object that) {
-        if (this == that) {
-            return true;
-        }
-        if (that == null) {
-            return false;
-        }
-        if (getClass() != that.getClass()) {
-            return false;
-        }
-        InterfaceEntity other = (InterfaceEntity) that;
-        return (this.getId() == null ? other.getId() == null : this.getId().equals(other.getId()))
-                && (this.getName() == null ? other.getName() == null : this.getName().equals(other.getName()))
-                && (this.getDescription() == null ? other.getDescription() == null : this.getDescription().equals(other.getDescription()))
-                && (this.getCallTimes() == null ? other.getCallTimes() == null : this.getCallTimes().equals(other.getCallTimes()))
-                && (this.getIsFree() == null ? other.getIsFree() == null : this.getIsFree().equals(other.getIsFree()))
-                && (this.getUrl() == null ? other.getUrl() == null : this.getUrl().equals(other.getUrl()))
-                && (this.getTag() == null ? other.getTag() == null : this.getTag().equals(other.getTag()))
-                && (this.getBackground() == null ? other.getBackground() == null : this.getBackground().equals(other.getBackground()))
-                && (this.getMethod() == null ? other.getMethod() == null : this.getMethod().equals(other.getMethod()))
-                && (this.getRequestParam() == null ? other.getRequestParam() == null : this.getRequestParam().equals(other.getRequestParam()))
-                && (this.getRequestHeader() == null ? other.getRequestHeader() == null : this.getRequestHeader().equals(other.getRequestHeader()))
-                && (this.getResponseHeader() == null ? other.getResponseHeader() == null : this.getResponseHeader().equals(other.getResponseHeader()))
-                && (this.getStatus() == null ? other.getStatus() == null : this.getStatus().equals(other.getStatus()))
-                && (this.getUserId() == null ? other.getUserId() == null : this.getUserId().equals(other.getUserId()))
-                && (this.getCreateTime() == null ? other.getCreateTime() == null : this.getCreateTime().equals(other.getCreateTime()))
-                && (this.getUpdateTime() == null ? other.getUpdateTime() == null : this.getUpdateTime().equals(other.getUpdateTime()))
-                && (this.getIsDelete() == null ? other.getIsDelete() == null : this.getIsDelete().equals(other.getIsDelete()));
-    }
-
-    @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = 1;
-        result = prime * result + ((getId() == null) ? 0 : getId().hashCode());
-        result = prime * result + ((getName() == null) ? 0 : getName().hashCode());
-        result = prime * result + ((getDescription() == null) ? 0 : getDescription().hashCode());
-        result = prime * result + ((getCallTimes() == null) ? 0 : getCallTimes().hashCode());
-        result = prime * result + ((getIsFree() == null) ? 0 : getIsFree().hashCode());
-        result = prime * result + ((getUrl() == null) ? 0 : getUrl().hashCode());
-        result = prime * result + ((getTag() == null) ? 0 : getTag().hashCode());
-        result = prime * result + ((getBackground() == null) ? 0 : getBackground().hashCode());
-        result = prime * result + ((getMethod() == null) ? 0 : getMethod().hashCode());
-        result = prime * result + ((getRequestParam() == null) ? 0 : getRequestParam().hashCode());
-        result = prime * result + ((getRequestHeader() == null) ? 0 : getRequestHeader().hashCode());
-        result = prime * result + ((getResponseHeader() == null) ? 0 : getResponseHeader().hashCode());
-        result = prime * result + ((getStatus() == null) ? 0 : getStatus().hashCode());
-        result = prime * result + ((getUserId() == null) ? 0 : getUserId().hashCode());
-        result = prime * result + ((getCreateTime() == null) ? 0 : getCreateTime().hashCode());
-        result = prime * result + ((getUpdateTime() == null) ? 0 : getUpdateTime().hashCode());
-        result = prime * result + ((getIsDelete() == null) ? 0 : getIsDelete().hashCode());
-        return result;
-    }
-
-    @Override
-    public String toString() {
-        StringBuilder sb = new StringBuilder();
-        sb.append(getClass().getSimpleName());
-        sb.append(" [");
-        sb.append("Hash = ").append(hashCode());
-        sb.append(", id=").append(id);
-        sb.append(", name=").append(name);
-        sb.append(", description=").append(description);
-        sb.append(", callTimes=").append(callTimes);
-        sb.append(", isFree=").append(isFree);
-        sb.append(", url=").append(url);
-        sb.append(", tag=").append(tag);
-        sb.append(", background=").append(background);
-        sb.append(", method=").append(method);
-        sb.append(", requestParam=").append(requestParam);
-        sb.append(", requestHeader=").append(requestHeader);
-        sb.append(", responseHeader=").append(responseHeader);
-        sb.append(", status=").append(status);
-        sb.append(", userId=").append(userId);
-        sb.append(", createTime=").append(createTime);
-        sb.append(", updateTime=").append(updateTime);
-        sb.append(", isDelete=").append(isDelete);
-        sb.append(", serialVersionUID=").append(serialVersionUID);
-        sb.append("]");
-        return sb.toString();
-    }
-}
